Tests/LibWeb/Text/expected/SVG/use-element-moveBefore-registry.txt

Use moved into shadow tree kept old fill: red
Use moved into document tree updated fill: green
